Consolidate extension classes into one place
