+++ src/GitHub.Exports/Extensions/VSExtensions.cs
