+++ src/GitHub.Extensions/GitRepoExtensions.cs
@@ -1,10 +1,6 @@
 ﻿using Microsoft.VisualStudio.TeamFoundation.Git.Extensibility;
 using NullGuard;
 using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace GitHub.Extensions
 {
